default group for ace admins
This allows server owners to become ESX admins on their first joins, based on if they have the `command` ace permission or not.  
If they do have that ace, they can `setgroup <id> admin`, so access-wise that's ok.  
I would commit this to master, but since v2 is not yet ready, PlumeESX uses "v1-final".