feat(es_extended/server/main.lua): Player Connect Logging
A new feature has been added, which allows for player identifiers to be captured and logged to a table (along with their last connect time), as well as the ability to send an embed to a Discord guild channel via ESX.DiscordLogFields. This can be disabled via a config option in config.lua with the webhook added into config.logs.lua which is mandatory if webhook logging is enabled.