fix(createJob): skin error
self.setJob in es_extended/server/classes/player.lua decodes skin data, which means skin data should be a string and not a table. When creating shops in run time and setting this job, it would throw error. 
Modified so it's a string and also added compatibility to parse skin_data directly into the job grade.